+++ src/ksi/list.h
@@ -19,11 +19,11 @@ int KSI_LIST_FN_NAME(type, append)(KSI_LIST(type) *, type *);					\
 int KSI_LIST_FN_NAME(type, iter)(KSI_LIST(type) *);								\
 int KSI_LIST_FN_NAME(type, next)(KSI_LIST(type) *, type **);					\
 int KSI_LIST_FN_NAME(type, indexOf)(const KSI_LIST(type) *, const type *);		\
-int KSI_LIST_FN_NAME(type, insertAt)(KSI_LIST(type) *, int, type *);			\
-int KSI_LIST_FN_NAME(type, replaceAt)(KSI_LIST(type) *, int, type *);			\
-int KSI_LIST_FN_NAME(type, remove)(KSI_LIST(type) *, int);						\
-int KSI_LIST_FN_NAME(type, length)(const KSI_LIST(type) *);						\
-int KSI_LIST_FN_NAME(type, elementAt)(const KSI_LIST(type) *, int pos, type **);		\
+int KSI_LIST_FN_NAME(type, insertAt)(KSI_LIST(type) *, size_t, type *);			\
+int KSI_LIST_FN_NAME(type, replaceAt)(KSI_LIST(type) *, size_t, type *);		\
+int KSI_LIST_FN_NAME(type, remove)(KSI_LIST(type) *, size_t);					\
+size_t KSI_LIST_FN_NAME(type, length)(const KSI_LIST(type) *);					\
+int KSI_LIST_FN_NAME(type, elementAt)(const KSI_LIST(type) *, size_t pos, type **);		\
 int KSI_LIST_FN_NAME(type, sort)(KSI_LIST(type) *, int (*)(const type *, const type *));	\
 KSI_CTX *type##List_getCtx(const KSI_LIST(type) *o);							\
 
@@ -33,14 +33,14 @@ void KSI_List_freeAll(KSI_List *list);
 void KSI_List_free(KSI_List *list);
 int KSI_List_new(void (*obj_free)(void *), KSI_List **list);
 int KSI_List_append(KSI_List *list, void *obj);
-int KSI_List_remove(KSI_List *list, unsigned int pos);
+int KSI_List_remove(KSI_List *list, size_t pos);
 int KSI_List_iter(KSI_List *list);
 int KSI_List_next(KSI_List *list, void **o);
 int KSI_List_indexOf(const KSI_List *list, const void *o);
-int KSI_List_insertAt(KSI_List *list, unsigned int pos, void *o);
-int KSI_List_replaceAt(KSI_List *list, unsigned int pos, void *o);
-int KSI_List_elementAt(const KSI_List *list, unsigned int pos, void **o);
-int KSI_List_length(const KSI_List *list);
+int KSI_List_insertAt(KSI_List *list, size_t pos, void *o);
+int KSI_List_replaceAt(KSI_List *list, size_t pos, void *o);
+int KSI_List_elementAt(const KSI_List *list, size_t pos, void **o);
+size_t KSI_List_length(const KSI_List *list);
 int KSI_List_sort(KSI_List *list, int (*)(const void *, const void *));
